# Add new `Performance/SortReverse` cop