@koic 2020-06-08
Yeah. I think that they should be separated into different cops of `CompareWithBlock`, `SortReverse`, and `RedundantSortBlock` because they have different roles.